refactor: rename folder ErrorBoundary
