+++ src/components/ErrorBoundary/error-boundary.scss
